added crc and sha256 options
Given a flash command, the users can now pass a sha256 hash of the given
image. This will be used by the BMC to verify the integrity of the
received data

`--skip_crc` can now be used to disable the crc readback step part of
the flash command. to be used at own risk